+++ src/App.js
@@ -3,7 +3,6 @@ import { Routes, Route } from 'react-router-dom';
 import Home from './components/pages/Home/Home';
 import Tienda from './components/pages/Tienda/Tienda';
 import ProfilePage from './components/pages/ProfilePage/ProfilePage';
-import PageFilt from './components/pageFilt/pageFilt';
 
 function App() {
   return (
