+++ src/components/Navbar/Navbar.jsx
@@ -1,22 +1,18 @@
 import React, { useState } from "react";
-import { Link, NavLink, useLocation } from "react-router-dom";
+import { Link, NavLink } from "react-router-dom";
 import "./NavBar.css";
 import { useAuth0 } from "@auth0/auth0-react";
 import LoginButton from "../LoginButton/LoginButton";
 import Logout from "../logout/logout";
 import Register from "../Register/Register";
-import SearchBar from "../SearchBar/SearchBar";
 
 function NavBar() {
   const { loginWithRedirect, logout, isAuthenticated } = useAuth0();
   const [click, setClick] = useState(false);
-  const location = useLocation();
 
   const handleClick = () => {
     setClick(!click);
   };
-
-  const isTiendaPage = location.pathname === "/tienda";
 
   return (
     <>
@@ -26,8 +22,6 @@ function NavBar() {
             <img className="img-hidden" src="/images/logo.PNG" alt="Logo" />
             <i className="fas fa-code"></i>
           </Link>
-
-          {isTiendaPage && <SearchBar className="search-bar" />}
 
           <ul className={click ? "nav-menu active" : "nav-menu"}>
             <li className="nav-item">
