--- src/components/SearchBar/SearchBar.css
@@ -1,47 +1,0 @@
-.search-bar-container {
-    position: absolute;
-    border: 1px solid;
-    width: 20%;
-    top: 4px;
-    left: 500px;
-    max-width: auto;
-    padding-top: 0.2rem;
-    padding-bottom: 0.5rem;
-    border-radius: 20px;
-    box-shadow: 0px 1px 2px rgb(0, 0, 0);
-    transition: box-shadow 0.3s;
-    cursor: pointer;
-  }
-  
-  .search-bar-wrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-  }
-  
-  .search-bar-label {
-    display: none;
-  }
-  
-  .search-bar-input {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 4rem;
-  }
-  
-  .search-bar-input input {
-    font-size: 0.875rem;
-    flex: 1;
-    outline: none;
-    background-color: transparent;
-    border: none;
-  }
-  
-  .search-bar-button {
-    padding: 0.5rem;
-    background-color: #f44336;
-    border-radius: 999px;
-    color: #ffffff;
-    cursor: pointer;
-  }
