+++ src/components/ToolBar/ToolBar.css
@@ -11,6 +11,7 @@
   justify-content: center;
   justify-content: space-around;
   flex-wrap: wrap;
+  margin-top: 40px;
 }
 
 .select {
@@ -43,4 +44,52 @@
 
 .toolbar-button:hover {
   background-color: #ddd;
+}
+
+
+.search-bar-container {
+  position: relative;
+  border: 1px solid;
+  width: 20%;
+  top: 30px;
+  left: 500px;
+  max-width: auto;
+  padding-top: 0.2rem;
+  padding-bottom: 0.5rem;
+  border-radius: 20px;
+  box-shadow: 0px 1px 2px rgb(0, 0, 0);
+  transition: box-shadow 0.3s;
+}
+
+.search-bar-wrapper {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.search-bar-label {
+  display: none;
+}
+
+.search-bar-input {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 4rem;
+}
+
+.search-bar-input input {
+  font-size: 0.875rem;
+  flex: 1;
+  outline: none;
+  background-color: transparent;
+  border: none;
+}
+
+.search-bar-button {
+  padding: 0.5rem;
+  background-color: #f44336;
+  border-radius: 999px;
+  color: #ffffff;
+  cursor: pointer;
 }
